pyproject.toml
[tool.poetry]
name = "biblioterra-scraper"
version = "0.1.0"
description = ""
authors = ["lamarcke <cassiolamarcksilvafreitas@gmail.com>"]
readme = "README.md"
packages = [{include = "scrapers_preservation"}]

[tool.poetry.dependencies]
python = "^3.10"
requests-html = "^0.10.0"
selenium = "^4.6.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
inquirer = "^2.10.1"
pyfiglet = "^0.8.post1"
yaspin = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
